+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/service/extra/hours/ExtraHoursService.java
@@ -53,4 +53,6 @@ public interface ExtraHoursService {
   public Message sendRefusalEmail(ExtraHours extraHours)
       throws AxelorException, ClassNotFoundException, InstantiationException,
           IllegalAccessException, MessagingException, IOException;
+
+  public void compute(ExtraHours extraHours);
 }